Take tour fragment out of the ecuation and fix menu bug in launcher
