+++ src/main/java/org/tndata/android/compass/fragment/LauncherFragment.java
@@ -38,7 +38,7 @@ public class LauncherFragment extends Fragment implements OnClickListener{
      */
     public LauncherFragment(){
         mAreViewsLoaded = false;
-        mShouldShowProgressOnLoad = false;
+        mShouldShowProgressOnLoad = true;
     }
 
     @Override
