+++ src/main/res/layout/activity_tour.xml
@@ -1,12 +1,45 @@
-<LinearLayout
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical"
     android:layout_width="match_parent"
-    android:layout_height="match_parent" >
+    android:layout_height="match_parent"
+    android:background="@color/tour_color" >
+
+    <org.tndata.android.compass.ui.SwipeOutViewPager
+        android:id="@+id/tour_pager"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_above="@+id/frameLayout" />
 
     <FrameLayout
-        android:id="@+id/my_priorities_fragment_host"
-        android:layout_width="match_parent"
-        android:layout_height="fill_parent" />
+        android:layout_width="fill_parent"
+        android:background="@android:color/white"
+        android:alpha="0.5"
+        android:focusable="false"
+        android:layout_height="1dp"
+        android:layout_above="@+id/tour_indicator"
+        android:layout_alignParentLeft="true"
+        android:layout_alignParentStart="true"
+        android:id="@+id/frameLayout" />
 
-</LinearLayout>
+    <me.relex.circleindicator.CircleIndicator
+        android:id="@+id/tour_indicator"
+        android:layout_width="wrap_content"
+        android:layout_height="40dp"
+        android:layout_alignParentBottom="true"
+        android:layout_centerHorizontal="true" />
+
+    <TextView
+        android:id="@+id/tour_skip"
+        android:layout_width="wrap_content"
+        android:layout_height="40dp"
+        android:layout_alignParentBottom="true"
+        android:layout_alignParentLeft="true"
+        android:layout_alignParentStart="true"
+        android:layout_toLeftOf="@+id/tour_indicator"
+        android:gravity="center"
+        android:text="@string/tour_action_skip"
+        android:textColor="@android:color/white"
+        android:textAllCaps="true" />
+
+</RelativeLayout>
