--- src/main/res/layout/fragment_tour.xml
@@ -1,51 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/tour_color"
-    android:orientation="vertical">
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="1">
-
-        <org.tndata.android.compass.ui.SwipeOutViewPager
-            android:id="@+id/tour_pager"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_above="@+id/frameLayout" />
-
-        <FrameLayout
-            android:layout_width="fill_parent"
-            android:background="@android:color/white"
-            android:alpha="0.5"
-            android:focusable="false"
-            android:layout_height="1dp"
-            android:layout_above="@+id/tour_indicator"
-            android:layout_alignParentLeft="true"
-            android:layout_alignParentStart="true"
-            android:id="@+id/frameLayout" />
-
-        <me.relex.circleindicator.CircleIndicator
-            android:id="@+id/tour_indicator"
-            android:layout_width="wrap_content"
-            android:layout_height="40dp"
-            android:layout_alignParentBottom="true"
-            android:layout_centerHorizontal="true" />
-
-        <TextView
-            android:id="@+id/tour_skip"
-            android:layout_width="wrap_content"
-            android:layout_height="40dp"
-            android:layout_alignParentBottom="true"
-            android:layout_alignParentLeft="true"
-            android:layout_alignParentStart="true"
-            android:layout_toLeftOf="@+id/tour_indicator"
-            android:gravity="center"
-            android:text="@string/tour_action_skip"
-            android:textColor="@android:color/white" />
-
-    </RelativeLayout>
-</LinearLayout>
